idex ticker("string") works in idexsmoketest
